requirements.txt
pandas 
openpyxl
python-dotenv 
streamlit 
openai
huggingface_hub
PyPDF2 
pyautogen
langchain 
spacey 
notebook
ipykernel
nltk
pypdf
chromadb 
tiktoken 
docx2txt
python-dotenv
markdown
typing_extensions
requests
'transformers[torch]'